requirements.txt
google-generativeai
django
django-rest-framework
